third_party/blink/web_tests/fast/dom/htmlcollection-protects-base-expected.txt

This test verifies that a HTMLCollection protects its base node from being GC'd.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS options[0].parentNode.getAttribute('foo') is 'bar'
PASS successfullyParsed is true

TEST COMPLETE
